refactor: move plugin database
