+++ packages/mask/background/database/plugin-db/wrap-plugin-database.ts
