+++ packages/mask/src/database/index.ts
@@ -1,5 +1,2 @@
-export * from '../../background/database/avatar-cache/avatar'
-export * from '../../background/database/post'
 export * from './Persona/helpers'
 export * from './Persona/types'
-export * from './Plugin'
