+++ packages/mask/src/plugins/RedPacket/Worker/index.ts
@@ -1,9 +1,12 @@
 import type { Plugin } from '@masknet/plugin-infra'
 import { base } from '../base'
 import '../messages'
+import { setupDatabase } from './database'
 
 const worker: Plugin.Worker.Definition = {
     ...base,
-    init(signal) {},
+    init(signal, context) {
+        setupDatabase(context.getDatabaseStorage())
+    },
 }
 export default worker
