+++ packages/mask/src/plugins/Wallet/Worker/index.ts
@@ -1,12 +1,14 @@
 import type { Plugin } from '@masknet/plugin-infra'
 import { base } from '@masknet/plugin-wallet'
 import { MaskMessages } from '../../../utils'
+import { setupDatabase } from '../database/Plugin.db'
 import '../messages'
 import { isLocked } from '../services'
 
 const worker: Plugin.Worker.Definition = {
     ...base,
-    init(signal) {
+    init(signal, context) {
+        setupDatabase(context.getDatabaseStorage())
         MaskMessages.events.wallet_is_locked.on(
             async ([type]) => {
                 if (type === 'request') {
